Switch release scripts to Javascript
Turns out that running the ts-node/esm loader is a HUGE memory hog, which is a real problem in production. Transpiling it first introduces a whole bunch of new hassles, so for a few scripts that prepare things for release, it makes sense to just write them in straight Javascript (still ES6, though).